backend/requirements.txt
fastapi[standard]
python-dotenv
openai
slowapi
